feat: product dto

+++ internal/domain/entity/product.go
@@ -8,7 +8,7 @@ type Product struct {
 	database.Model
 	Name        string     `gorm:"type:varchar(100);not null;index"`
 	Description string     `gorm:"type:text"`
-	Price       float64    `gorm:"not null"`
+	Price       int        `gorm:"not null"`
 	Inventory   int        `gorm:"not null"`
 	CategoryID  uint       `gorm:"index"`
 	Category    Category   `gorm:"constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
